# server_swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/libraries/{id}": {
      "get": {
        "operationId": "LibrariesController_findOneByIdWithPermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          }
        },
        "tags": [
          "libraries"
        ]
      }
    },
    "/v1/collections/{id}": {
      "get": {
        "operationId": "CollectionsController_findOneByIdWithPermissions",
        "summary": "",
        "description": "Returns a collection which the user has access to\n\n(Either its own collection or a public one)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          }
        },
        "tags": [
          "collections"
        ]
      },
      "patch": {
        "operationId": "CollectionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "collections"
        ]
      }
    },
    "/v1/collections": {
      "post": {
        "operationId": "CollectionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "collections"
        ]
      }
    },
    "/v1/reviews": {
      "post": {
        "operationId": "ReviewsController_createOrUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "reviews"
        ]
      },
      "get": {
        "operationId": "ReviewsController_findOneByUserIdAndGameId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
    "/v1/reviews/game/{id}": {
      "get": {
        "operationId": "ReviewsController_findAllByGameId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindReviewPaginatedDto"
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
    "/v1/reviews/{id}": {
      "get": {
        "operationId": "ReviewsController_findOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ]
      },
      "delete": {
        "operationId": "ReviewsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "reviews"
        ]
      }
    },
...
    "/v1/collections-entries": {
      "post": {
        "operationId": "CollectionsEntriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "collections-entries"
        ]
      }
    },
    "/v1/collections-entries/game/{id}": {
      "get": {
        "operationId": "CollectionsEntriesController_findOwnEntryByGameId",
        "summary": "",
        "description": "Returns a specific collection entry based on game ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query"
          }
        },
        "tags": [
          "collections-entries"
        ]
      },
      "delete": {
        "operationId": "CollectionsEntriesController_deleteOwnEntryByGameId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "collections-entries"
        ]
      }
    },
    "/v1/collections-entries/game/{id}/favorite": {
      "post": {
        "operationId": "CollectionsEntriesController_changeFavoriteStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteStatusCollectionEntryDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "collections-entries"
        ]
      }
    },
    "/v1/collections-entries/collection/{id}": {
      "get": {
        "operationId": "CollectionsEntriesController_findAllByCollectionId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionEntriesPaginatedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "collections-entries"
        ]
      }
    },
...
    "/v1/game/repository/{id}": {
      "post": {
        "operationId": "GameRepositoryController_findOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameRepositoryFindOneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          }
        },
        "tags": [
          "game-repository"
        ]
      }
    },
    "/v1/game/repository": {
      "post": {
        "operationId": "GameRepositoryController_findAllByIds",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameRepositoryFindAllDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameRepositoryPaginatedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "game-repository"
        ]
      }
    },
...
  },
  "info": {
    "title": "GameNode API",
    "description": "API docs for the videogame catalog system GameNode. <br><br>Built with love by the GameNode team.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Library": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Also used to share the library with other users.\n\nSame as SuperTokens' userId."
          },
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Collection"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "collections",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "GamePlatform": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "abbreviation": {
            "type": "string"
          },
          "alternative_name": {
            "type": "string"
          },
          "category": {
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ],
            "type": "number"
          },
          "checksum": {
            "type": "string"
          },
          "generation": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Game"
            }
          },
          "collectionEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionEntry"
            }
          }
        },
        "required": [
          "id",
          "abbreviation",
          "alternative_name",
          "category",
          "checksum",
          "generation",
          "name",
          "createdAt",
          "updatedAt",
          "games",
          "collectionEntries"
        ]
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Shareable string ID\n\nSame as SuperTokens' userId."
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "$ref": "#/components/schemas/ProfileAvatar"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "username",
          "avatar",
          "createdAt",
          "updatedAt"
        ]
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "collectionEntry": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionEntry"
            }
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "rating",
          "game",
          "collectionEntry",
          "profile",
          "createdAt",
          "updatedAt"
        ]
      },
      "CollectionEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "review": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Review"
              }
            ]
          },
          "reviewId": {
            "type": "string",
            "nullable": true
          },
          "collection": {
            "$ref": "#/components/schemas/Collection"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "gameId": {
            "type": "number"
          },
          "ownedPlatforms": {
            "description": "The platforms on which the user owns the game.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GamePlatform"
            }
          },
          "isFavorite": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reviewId",
          "collection",
          "game",
          "gameId",
          "ownedPlatforms",
          "isFavorite",
          "createdAt",
          "updatedAt"
        ]
      },
      "Collection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "library": {
            "$ref": "#/components/schemas/Library"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionEntry"
            }
          },
          "isFeatured": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "isPublic",
          "library",
          "entries",
          "isFeatured",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "default": true
          },
          "isFeatured": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "isPublic",
          "isFeatured"
        ]
      },
      "UpdateCollectionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "number"
          },
          "content": {
            "type": "string",
            "minLength": 3
          },
          "rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          }
        },
        "required": [
          "gameId",
          "content",
          "rating"
        ]
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "number",
            "description": "Total number of items available for the current query"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages available for the current query"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "If this query allows for a next page"
          }
        },
        "required": [
          "totalItems",
          "totalPages",
          "hasNextPage"
        ]
      },
      "FindReviewPaginatedDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
...
      "CreateCollectionEntryDto": {
        "type": "object",
        "properties": {
          "collectionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gameId": {
            "type": "number"
          },
          "platformIds": {
            "type": "array",
            "items": {
              "type": "number",
              "enum": [
                6,
                7,
                8,
                9,
                48,
                167,
                11,
                12,
                49,
                169,
                130,
                170
              ]
            }
          },
          "isFavorite": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "collectionIds",
          "gameId",
          "platformIds",
          "isFavorite"
        ]
      },
      "FavoriteStatusCollectionEntryDto": {
        "type": "object",
        "properties": {
          "isFavorite": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "isFavorite"
        ]
      },
      "CollectionEntriesPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionEntry"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
...
      "GameRepositoryFindOneDto": {
        "type": "object",
        "properties": {
          "relations": {
            "type": "object"
          }
        }
      },
      "GameRepositoryFindAllDto": {
        "type": "object",
        "properties": {
          "gameIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "relations": {
            "type": "object"
          },
          "search": {
            "type": "string"
          },
          "offset": {
            "type": "number",
            "default": 0
          },
          "limit": {
            "type": "number",
            "default": 20
          },
          "orderBy": {
            "type": "object"
          }
        },
        "required": [
          "gameIds"
        ]
      },
      "GameRepositoryPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Game"
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}